requirements.txt
Flask==1.1.2

line-bot-sdk==1.8.0